[IMP] Implements technical sidebar
It introduces a sidebar (toggleable with a button) that displays DB information and the fields available in the view.
When you hover over a field in the list, it highlights it in the view and scrolls to its position if it is hidden.

There is also a selector mode where you can select a specific field and get its information.

It also gives you some information when you are on the website (it also works in editor mode).

The sidebar can be activated or deactivated through the options page, but also via the popup for a quick switch.

It introduces an update popup to let users know when there is an important update.

Remove some useEffect hooks that were not necessary.